if subrows are disabled then don't allow them to be expanded
git-svn-id: http://svn.vuze.com/public/client@40677 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/ui/swt/views/table/painted/TableRowPainted.java
@@ -827,42 +827,44 @@ public class TableRowPainted
 
 	@Override
 	public void setExpanded(boolean b) {
-		int oldHeight = getFullHeight();
-		super.setExpanded(b);
-		synchronized (subRows_sync) {
-			TableRowPainted[] newSubRows = null;
-			if (b && (subRows == null || subRows.length != numSubItems)
-					&& subDataSources != null && subDataSources.length == numSubItems) {
-				if (DEBUG_SUBS) {
-					debug("building subrows " + numSubItems);
+		if ( canExpand() ){
+			int oldHeight = getFullHeight();
+			super.setExpanded(b);
+			synchronized (subRows_sync) {
+				TableRowPainted[] newSubRows = null;
+				if (b && (subRows == null || subRows.length != numSubItems)
+						&& subDataSources != null && subDataSources.length == numSubItems) {
+					if (DEBUG_SUBS) {
+						debug("building subrows " + numSubItems);
+					}
+	
+					deleteExistingSubRows();
+					newSubRows = new TableRowPainted[numSubItems];
+					TableViewPainted tv = getViewPainted();
+					int h = 0;
+					for (int i = 0; i < newSubRows.length; i++) {
+						newSubRows[i] = new TableRowPainted(this, tv, subDataSources[i],
+								false);
+						newSubRows[i].setTableItem(i, false);
+						h += newSubRows[i].getHeight();
+					}
+	
+					subRowsHeight = h;
+	
+					subRows = newSubRows;
 				}
-
-				deleteExistingSubRows();
-				newSubRows = new TableRowPainted[numSubItems];
-				TableViewPainted tv = getViewPainted();
-				int h = 0;
-				for (int i = 0; i < newSubRows.length; i++) {
-					newSubRows[i] = new TableRowPainted(this, tv, subDataSources[i],
-							false);
-					newSubRows[i].setTableItem(i, false);
-					h += newSubRows[i].getHeight();
+	
+				getViewPainted().rowHeightChanged(this, oldHeight, getFullHeight());
+	
+				if (newSubRows != null) {
+					getViewPainted().triggerListenerRowAdded(newSubRows);
 				}
-
-				subRowsHeight = h;
-
-				subRows = newSubRows;
+	
 			}
-
-			getViewPainted().rowHeightChanged(this, oldHeight, getFullHeight());
-
-			if (newSubRows != null) {
-				getViewPainted().triggerListenerRowAdded(newSubRows);
+			if (isVisible()) {
+				getViewPainted().visibleRowsChanged();
+				getViewPainted().redrawTable();
 			}
-
-		}
-		if (isVisible()) {
-			getViewPainted().visibleRowsChanged();
-			getViewPainted().redrawTable();
 		}
 	}
